transport: Fix _find_myself for kernel 6.9
Kernel 6.9 seems to have changed AF_NETLINK behavior slightly making
booth unable to start.

Previously it was expected only first item in
the message can be NLMSG_DONE or NLMSG_ERROR type. And it looks this was
true for Kernel < 6.9.

With kernel 6.9 this is no longer true, so any item can be type
NLMSG_DONE or NLMSG_ERROR.

Result was loop was never terminated and booth was waiting for more
messages from kernel which never arrived.

Solution is to change loop a bit so NLMSG_DONE, NLMSG_ERROR and
RTM_NEWADDR are handled correctly.

Signed-off-by: Jan Friesse <jfriesse@redhat.com>